Extracted the faraday feature to a faraday provider
This is the base for realising a ruby provider without faraday. But use the faraday as a fallback. It also removed the `feature => faraday` for the other providers
