+++ lib/puppet/provider/archive/ruby.rb
@@ -1,12 +1,10 @@
 begin
   require 'puppet_x/bodeco/archive'
-  require 'puppet_x/bodeco/util'
 rescue LoadError
   require 'pathname' # WORK_AROUND #14073 and #7788
   archive = Puppet::Module.find('archive', Puppet[:environment].to_s)
   raise(LoadError, "Unable to find archive module in modulepath #{Puppet[:basemodulepath] || Puppet[:modulepath]}") unless archive
   require File.join archive.path, 'lib/puppet_x/bodeco/archive'
-  require File.join archive.path, 'lib/puppet_x/bodeco/util'
 end
 
 require 'securerandom'
@@ -14,6 +12,8 @@ require 'tempfile'
 
 Puppet::Type.type(:archive).provide(:ruby) do
   attr_reader :archive_checksum
+
+  confine :true => false # This is NEVER a valid provider. It is just used as a base class
 
   def exists?
     if extracted?
@@ -48,23 +48,6 @@ Puppet::Type.type(:archive).provide(:ruby) do
     else
       "#{resource[:filename]}_#{resource[:checksum]}"
     end
-  end
-
-  def download(archive_filepath)
-    tempfile = Tempfile.new(tempfile_name)
-    temppath = tempfile.path
-    tempfile.close!
-
-    PuppetX::Bodeco::Util.download(resource[:source], temppath, :username => resource[:username], :password => resource[:password], :cookie => resource[:cookie] )
-
-    # conditionally verify checksum:
-    if resource[:checksum_verify] == :true and resource[:checksum_type] != :none
-
-      archive = PuppetX::Bodeco::Archive.new(temppath)
-      raise(Puppet::Error, 'Download file checksum mismatch') unless archive.checksum(resource[:checksum_type]) == checksum
-    end
-
-    FileUtils.mv(temppath, archive_filepath)
   end
 
   def creates
