+++ spec/unit/puppet/provider/archive/faraday_spec.rb
@@ -1,0 +1,9 @@
+require 'spec_helper'
+require 'tmpdir'
+
+faraday_provider = Puppet::Type.type(:archive).provider(:faraday)
+
+
+RSpec.describe faraday_provider do
+  it_behaves_like 'an archive provider', faraday_provider
+end
